requirements.txt
torch>=1.7.0
torchvision>=0.8.1
numpy>=1.18.5
opencv-python>=4.1.2
Pillow>=7.1.2
pyYAML>=5.3.1
tqdm>=4.41.0
matplotlib>=3.2.2
seaborn>=0.11.0
pandas>=1.1.4
albumentations>=0.5.2